refactor scene ui 3
