+++ assets/ui/combo.dui
@@ -1,0 +1,21 @@
+<define-template id="combo-root">
+    <div style="width: 100%; min-width: 100%; border: 1px; border-color: #88888888;">
+        <div id="selection" style="flex-grow: 1;">
+            <div id="text" text="@selection" />
+            <space />
+            <div>
+                <bounds style="height: 80%; width: auto; aspect-ratio: 1.0; top: 10%; right: 10%" bound-image="images/left-arrow.png" />
+            </div>
+        </div>
+    </div>
+</define-template>
+
+<define-template id="combo-popup">
+    <div style="position-type: absolute; top: '@top'; left: '@left'; height: '@height'; min-width: '@width'; border: 1px; border-color: #88888888; background-color: #000000dd" z-index="100">
+        <div style="flex-direction: column; width: 100%; height: 100%">
+            <vscroll>
+                <div id="contents" style="width: 100%; flex-direction: column; margin: 0px 2vmin 0px 0px;" />
+            </vscroll>
+        </div>
+    </div>
+</define-template>
