+++ crates/scene_runner/src/update_world/scene_ui/mod.rs
@@ -1,3 +1,9 @@
+pub mod ui_background;
+pub mod ui_dropdown;
+pub mod ui_input;
+pub mod ui_pointer;
+pub mod ui_text;
+
 use std::collections::{BTreeMap, BTreeSet};
 
 use bevy::{
@@ -8,43 +14,37 @@ use bevy::{
     utils::{HashMap, HashSet},
 };
 use bevy_dui::{DuiCommandsExt, DuiProps, DuiRegistry};
+use ui_background::{set_ui_background, UiBackground};
+use ui_dropdown::{set_ui_dropdown, UiDropdown};
+use ui_input::{set_ui_input, UiInput};
+use ui_pointer::set_ui_pointer_events;
+use ui_text::{set_ui_text, UiText};
 
 use crate::{
-    renderer_context::RendererSceneContext, update_scene::pointer_results::UiPointerTarget,
-    update_world::text_shape::make_text_section, ContainerEntity, ContainingScene, SceneEntity,
+    renderer_context::RendererSceneContext, ContainerEntity, ContainingScene, SceneEntity,
     SceneSets,
 };
 use common::{
     structs::{AppConfig, PrimaryUser},
-    util::{DespawnWith, FireEventEx, ModifyComponentExt, TryPushChildrenEx},
+    util::{DespawnWith, FireEventEx, ModifyComponentExt},
 };
 use dcl::interface::{ComponentPosition, CrdtType};
 use dcl_component::{
-    proto_components::{
-        self,
-        common::{texture_union, BorderRect, TextureUnion},
-        sdk::components::{
-            self, scroll_position_value, PbUiBackground, PbUiCanvas, PbUiDropdown,
-            PbUiDropdownResult, PbUiInput, PbUiInputResult, PbUiScrollResult, PbUiText,
-            PbUiTransform, ScrollPositionValue, TextWrap, YgAlign, YgDisplay, YgFlexDirection,
-            YgJustify, YgOverflow, YgPositionType, YgUnit, YgWrap,
-        },
+    proto_components::sdk::components::{
+        self, scroll_position_value, PbUiBackground, PbUiCanvas, PbUiDropdown, PbUiInput,
+        PbUiScrollResult, PbUiText, PbUiTransform, ScrollPositionValue, YgAlign, YgDisplay,
+        YgFlexDirection, YgJustify, YgOverflow, YgPositionType, YgUnit, YgWrap,
     },
     SceneComponentId, SceneEntityId,
 };
 use ui_core::{
-    combo_box::ComboBox,
-    nine_slice::Ui9Slice,
     scrollable::{
         ScrollDirection, ScrollPosition, ScrollTarget, ScrollTargetEvent, Scrollable, StartPosition,
     },
-    stretch_uvs_image::StretchUvMaterial,
-    textentry::TextEntry,
-    ui_actions::{DataChanged, HoverEnter, HoverExit, On, Submit, UiCaller},
-    ui_builder::SpawnSpacer,
+    ui_actions::{DataChanged, On, UiCaller},
 };
 
-use super::{material::TextureResolver, pointer_events::PointerEvents, AddCrdtInterfaceExt};
+use super::AddCrdtInterfaceExt;
 
 pub struct SceneUiPlugin;
 
@@ -285,156 +285,11 @@ impl From<PbUiTransform> for UiTransform {
     }
 }
 
-#[derive(Clone, Copy, Debug)]
-pub enum BackgroundTextureMode {
-    NineSlices(BorderRect),
-    Stretch([Vec4; 2]),
-    Center,
-}
-
-impl BackgroundTextureMode {
-    pub fn stretch_default() -> Self {
-        Self::Stretch([Vec4::W, Vec4::ONE - Vec4::W])
-    }
-}
-
-#[derive(Component, Clone, Debug)]
-pub struct BackgroundTexture {
-    tex: TextureUnion,
-    mode: BackgroundTextureMode,
-}
-
-#[derive(Component, Clone, Debug)]
-pub struct UiBackground {
-    color: Option<Color>,
-    texture: Option<BackgroundTexture>,
-}
-
-impl From<PbUiBackground> for UiBackground {
-    fn from(value: PbUiBackground) -> Self {
-        let texture_mode = value.texture_mode();
-        Self {
-            color: value.color.map(Into::into),
-            texture: value.texture.map(|tex| {
-                let mode = match texture_mode {
-                    components::BackgroundTextureMode::NineSlices => {
-                        BackgroundTextureMode::NineSlices(value.texture_slices.unwrap_or(
-                            BorderRect {
-                                top: 1.0 / 3.0,
-                                bottom: 1.0 / 3.0,
-                                left: 1.0 / 3.0,
-                                right: 1.0 / 3.0,
-                            },
-                        ))
-                    }
-                    components::BackgroundTextureMode::Center => BackgroundTextureMode::Center,
-                    components::BackgroundTextureMode::Stretch => {
-                        // the uvs array contain [tl.x, tl.y, bl.x, bl.y, br.x, br.y, tr.x, tr.y]
-                        let mut iter = value.uvs.iter().copied();
-                        let uvs = [
-                            Vec4::new(
-                                iter.next().unwrap_or(0.0),
-                                iter.next().unwrap_or(0.0),
-                                iter.next().unwrap_or(0.0),
-                                iter.next().unwrap_or(1.0),
-                            ),
-                            Vec4::new(
-                                iter.next().unwrap_or(1.0),
-                                iter.next().unwrap_or(1.0),
-                                iter.next().unwrap_or(1.0),
-                                iter.next().unwrap_or(0.0),
-                            ),
-                        ];
-                        BackgroundTextureMode::Stretch(uvs)
-                    }
-                };
-
-                BackgroundTexture { tex, mode }
-            }),
-        }
-    }
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum VAlign {
-    Top,
-    Middle,
-    Bottom,
-}
-
-#[derive(Component, Clone, Debug)]
-pub struct UiText {
-    pub text: String,
-    pub color: Color,
-    pub h_align: JustifyText,
-    pub v_align: VAlign,
-    pub font: proto_components::sdk::components::common::Font,
-    pub font_size: f32,
-    pub wrapping: bool,
-}
-
-impl From<PbUiText> for UiText {
-    fn from(value: PbUiText) -> Self {
-        let text_align = value
-            .text_align
-            .map(|_| value.text_align())
-            .unwrap_or(components::common::TextAlignMode::TamMiddleCenter);
-
-        Self {
-            text: value.value.clone(),
-            color: value.color.map(Into::into).unwrap_or(Color::WHITE),
-            h_align: match text_align {
-                components::common::TextAlignMode::TamTopLeft
-                | components::common::TextAlignMode::TamMiddleLeft
-                | components::common::TextAlignMode::TamBottomLeft => JustifyText::Left,
-                components::common::TextAlignMode::TamTopCenter
-                | components::common::TextAlignMode::TamMiddleCenter
-                | components::common::TextAlignMode::TamBottomCenter => JustifyText::Center,
-                components::common::TextAlignMode::TamTopRight
-                | components::common::TextAlignMode::TamMiddleRight
-                | components::common::TextAlignMode::TamBottomRight => JustifyText::Right,
-            },
-            v_align: match text_align {
-                components::common::TextAlignMode::TamTopLeft
-                | components::common::TextAlignMode::TamTopCenter
-                | components::common::TextAlignMode::TamTopRight => VAlign::Top,
-                components::common::TextAlignMode::TamMiddleLeft
-                | components::common::TextAlignMode::TamMiddleCenter
-                | components::common::TextAlignMode::TamMiddleRight => VAlign::Middle,
-                components::common::TextAlignMode::TamBottomLeft
-                | components::common::TextAlignMode::TamBottomCenter
-                | components::common::TextAlignMode::TamBottomRight => VAlign::Bottom,
-            },
-            font: value.font(),
-            font_size: value.font_size.unwrap_or(10) as f32,
-            wrapping: value.text_wrap() == TextWrap::TwWrap,
-        }
-    }
-}
-
 #[derive(Component, Clone)]
 pub struct UiCanvas(pub PbUiCanvas);
 
 impl From<PbUiCanvas> for UiCanvas {
     fn from(value: PbUiCanvas) -> Self {
-        Self(value)
-    }
-}
-
-#[derive(Component, Debug)]
-pub struct UiInput(PbUiInput);
-
-impl From<PbUiInput> for UiInput {
-    fn from(value: PbUiInput) -> Self {
-        Self(value)
-    }
-}
-
-#[derive(Component, Debug)]
-pub struct UiDropdown(PbUiDropdown);
-
-impl From<PbUiDropdown> for UiDropdown {
-    fn from(value: PbUiDropdown) -> Self {
         Self(value)
     }
 }
@@ -1079,542 +934,6 @@ fn layout_scene_ui(
             } else {
                 warn!("scroll to target `{target}` not found");
             }
-        }
-    }
-}
-
-fn set_ui_input(
-    mut commands: Commands,
-    inputs: Query<(&SceneEntity, &UiInput, &UiLink), Or<(Changed<UiInput>, Changed<UiLink>)>>,
-    mut removed: RemovedComponents<UiInput>,
-    links: Query<&UiLink>,
-) {
-    for ent in removed.read() {
-        if let Ok(link) = links.get(ent) {
-            if let Some(mut commands) = commands.get_entity(link.ui_entity) {
-                commands.remove::<TextEntry>();
-            }
-        }
-    }
-
-    for (scene_ent, input, link) in inputs.iter() {
-        let Some(mut commands) = commands.get_entity(link.ui_entity) else {
-            continue;
-        };
-
-        let font_size = input.0.font_size.unwrap_or(12);
-        let ui_entity = link.ui_entity;
-        let root = scene_ent.root;
-        let scene_id = scene_ent.id;
-
-        let data_handler = move |In(submit): In<bool>,
-                                 entry: Query<&TextEntry>,
-                                 mut context: Query<&mut RendererSceneContext>,
-                                 time: Res<Time>,
-                                 caller: Res<UiCaller>| {
-            debug!("callback on {:?}", caller.0);
-            let Ok(entry) = entry.get(ui_entity) else {
-                warn!("failed to get text node on UiInput update");
-                return;
-            };
-            let Ok(mut context) = context.get_mut(root) else {
-                warn!("failed to get context on UiInput update");
-                return;
-            };
-
-            context.update_crdt(
-                SceneComponentId::UI_INPUT_RESULT,
-                CrdtType::LWW_ENT,
-                scene_id,
-                &PbUiInputResult {
-                    value: entry.content.clone(),
-                    is_submit: Some(submit),
-                },
-            );
-            context.last_action_event = Some(time.elapsed_seconds());
-        };
-
-        commands.modify_component(move |style: &mut Style| {
-            //ensure we use max width if not given
-            if style.width == Val::Px(0.0) {
-                style.width = Val::Percent(100.0);
-            }
-            //and some size if not given
-            if style.height == Val::Px(0.0) {
-                style.height = Val::Px(font_size as f32 * 1.3);
-            }
-        });
-
-        commands.try_insert((
-            FocusPolicy::Block,
-            Interaction::default(),
-            TextEntry {
-                hint_text: input.0.placeholder.to_owned(),
-                enabled: !input.0.disabled,
-                content: input.0.value.clone().unwrap_or_default(),
-                accept_line: false,
-                font_size,
-                ..Default::default()
-            },
-            On::<DataChanged>::new((|| false).pipe(data_handler)),
-            On::<Submit>::new((|| true).pipe(data_handler)),
-        ));
-    }
-}
-
-fn set_ui_dropdown(
-    mut commands: Commands,
-    dropdowns: Query<
-        (&SceneEntity, &UiDropdown, &UiLink),
-        Or<(Changed<UiDropdown>, Changed<UiLink>)>,
-    >,
-    mut removed: RemovedComponents<UiDropdown>,
-    links: Query<&UiLink>,
-) {
-    for ent in removed.read() {
-        if let Ok(link) = links.get(ent) {
-            if let Some(mut commands) = commands.get_entity(link.ui_entity) {
-                commands.remove::<ComboBox>();
-            }
-        }
-    }
-
-    for (scene_ent, dropdown, link) in dropdowns.iter() {
-        let Some(mut commands) = commands.get_entity(link.ui_entity) else {
-            continue;
-        };
-
-        let initial_selection = if dropdown.0.accept_empty {
-            dropdown.0.selected_index.map(|ix| ix as isize)
-        } else {
-            Some(dropdown.0.selected_index.unwrap_or(0) as isize)
-        };
-
-        commands.modify_component(|style: &mut Style| {
-            //ensure we use max width if not given
-            if style.width == Val::Px(0.0) || style.width == Val::Auto {
-                style.width = Val::Percent(100.0);
-            }
-            //and some size if not given
-            if style.height == Val::Px(0.0) || style.height == Val::Auto {
-                style.height = Val::Px(16.0);
-            }
-        });
-
-        let root = scene_ent.root;
-        let ui_entity = link.ui_entity;
-        let scene_id = scene_ent.id;
-        commands.try_insert((
-            ComboBox::new(
-                dropdown.0.empty_label.clone().unwrap_or_default(),
-                &dropdown.0.options,
-                dropdown.0.accept_empty,
-                dropdown.0.disabled,
-                initial_selection,
-            ),
-            On::<DataChanged>::new(
-                move |combo: Query<(Entity, &ComboBox)>,
-                      mut context: Query<&mut RendererSceneContext>,
-                      time: Res<Time>| {
-                    let Ok((_, combo)) = combo.get(ui_entity) else {
-                        warn!("failed to get combo node on UiDropdown update");
-                        return;
-                    };
-                    let Ok(mut context) = context.get_mut(root) else {
-                        warn!("failed to get context on UiInput update");
-                        return;
-                    };
-
-                    context.update_crdt(
-                        SceneComponentId::UI_DROPDOWN_RESULT,
-                        CrdtType::LWW_ENT,
-                        scene_id,
-                        &PbUiDropdownResult {
-                            value: combo.selected as i32,
-                        },
-                    );
-                    context.last_action_event = Some(time.elapsed_seconds());
-                },
-            ),
-        ));
-    }
-}
-
-fn set_ui_pointer_events(
-    mut commands: Commands,
-    pes: Query<
-        (Entity, &UiLink),
-        (
-            With<PointerEvents>,
-            Or<(Changed<PointerEvents>, Changed<UiLink>)>,
-        ),
-    >,
-) {
-    for (ent, link) in pes.iter() {
-        if let Some(mut commands) = commands.get_entity(link.ui_entity) {
-            let is_primary = link.is_window_ui;
-            commands.try_insert((
-                FocusPolicy::Block,
-                Interaction::default(),
-                On::<HoverEnter>::new(move |mut ui_target: ResMut<UiPointerTarget>| {
-                    if is_primary {
-                        *ui_target = UiPointerTarget::Primary(ent);
-                    } else {
-                        *ui_target = UiPointerTarget::World(ent);
-                    }
-                }),
-                On::<HoverExit>::new(move |mut ui_target: ResMut<UiPointerTarget>| {
-                    if *ui_target == UiPointerTarget::Primary(ent)
-                        || *ui_target == UiPointerTarget::World(ent)
-                    {
-                        *ui_target = UiPointerTarget::None;
-                    };
-                }),
-            ));
-        }
-    }
-}
-
-#[derive(Component)]
-pub struct UiTextMarker;
-
-fn set_ui_text(
-    mut commands: Commands,
-    texts: Query<(&UiText, &UiTransform, &UiLink), Or<(Changed<UiText>, Changed<UiLink>)>>,
-    mut removed: RemovedComponents<UiText>,
-    links: Query<&UiLink>,
-    children: Query<&Children>,
-    prev_texts: Query<&UiTextMarker>,
-) {
-    for ent in removed.read() {
-        let Ok(link) = links.get(ent) else {
-            continue;
-        };
-
-        if let Ok(children) = children.get(link.ui_entity) {
-            for child in children.iter().filter(|c| prev_texts.get(**c).is_ok()) {
-                if let Some(commands) = commands.get_entity(*child) {
-                    commands.despawn_recursive();
-                }
-            }
-        }
-    }
-
-    for (ui_text, ui_transform, link) in texts.iter() {
-        // remove old text
-        if let Ok(children) = children.get(link.ui_entity) {
-            for child in children.iter().filter(|c| prev_texts.get(**c).is_ok()) {
-                if let Some(commands) = commands.get_entity(*child) {
-                    commands.despawn_recursive();
-                }
-            }
-        }
-
-        if ui_text.text.is_empty() || ui_text.font_size <= 0.0 {
-            continue;
-        }
-
-        let Some(mut ent_cmds) = commands.get_entity(link.ui_entity) else {
-            continue;
-        };
-
-        let text = make_text_section(
-            ui_text.text.as_str(),
-            ui_text.font_size,
-            ui_text
-                .color
-                .with_alpha(ui_text.color.alpha() * link.opacity.0),
-            ui_text.font,
-            ui_text.h_align,
-            ui_text.wrapping,
-        );
-
-        // with text nodes the axis sizes are unusual.
-        // a) if either size axis is NOT NONE, (explicit or auto), we want auto to size appropriately for the content.
-        // b) if both axes are NONE, we want to size to zero.
-        // a) - we tackle this by using a nested position-type: relative node which will size it's parent appropriately, and default the parent to Auto
-        //    - for alignment we use align-items and justify-content
-        // b) - we use a nested position-type: absolute node, and default the parent to auto
-        //    - for alignment we use align-items and justify-content as above, and we also set left/right/top/bottom to 50% if required
-
-        let any_axis_specified = [ui_transform.size.width, ui_transform.size.height]
-            .iter()
-            .any(Option::is_some);
-
-        let inner_style = if any_axis_specified {
-            Style {
-                position_type: PositionType::Relative,
-                ..Default::default()
-            }
-        } else {
-            Style {
-                position_type: PositionType::Absolute,
-                left: if ui_text.h_align == JustifyText::Left {
-                    Val::Percent(50.0)
-                } else {
-                    Val::Auto
-                },
-                right: if ui_text.h_align == JustifyText::Right {
-                    Val::Percent(50.0)
-                } else {
-                    Val::Auto
-                },
-                top: if ui_text.v_align == VAlign::Top {
-                    Val::Percent(50.0)
-                } else {
-                    Val::Auto
-                },
-                bottom: if ui_text.v_align == VAlign::Bottom {
-                    Val::Percent(50.0)
-                } else {
-                    Val::Auto
-                },
-                ..Default::default()
-            }
-        };
-
-        // we need to set size for the first inner element depending
-        // on how the outer was specified
-        let width = match ui_transform.size.width {
-            Some(Val::Px(px)) => Val::Px(px),
-            Some(Val::Percent(_)) => Val::Percent(100.0),
-            _ => Val::Auto,
-        };
-        let height = match ui_transform.size.height {
-            Some(Val::Px(px)) => Val::Px(px),
-            Some(Val::Percent(_)) => Val::Percent(100.0),
-            _ => Val::Auto,
-        };
-
-        ent_cmds.try_with_children(|c| {
-            c.spawn((
-                NodeBundle {
-                    style: Style {
-                        flex_direction: FlexDirection::Row,
-                        justify_content: match ui_text.h_align {
-                            JustifyText::Left => JustifyContent::FlexStart,
-                            JustifyText::Center => JustifyContent::Center,
-                            JustifyText::Right => JustifyContent::FlexEnd,
-                            JustifyText::Justified => unreachable!(),
-                        },
-                        align_items: match ui_text.v_align {
-                            VAlign::Top => AlignItems::FlexStart,
-                            VAlign::Middle => AlignItems::Center,
-                            VAlign::Bottom => AlignItems::FlexEnd,
-                        },
-                        width,
-                        height,
-                        align_self: AlignSelf::FlexStart,
-                        // elements are horizontally centered by default
-                        margin: UiRect::horizontal(Val::Auto),
-                        ..Default::default()
-                    },
-                    ..Default::default()
-                },
-                UiTextMarker,
-            ))
-            .try_with_children(|c| {
-                c.spawn(NodeBundle {
-                    style: inner_style,
-                    ..Default::default()
-                })
-                .try_with_children(|c| {
-                    c.spawn(TextBundle {
-                        text,
-                        z_index: ZIndex::Local(1),
-                        ..Default::default()
-                    });
-                });
-            });
-        });
-    }
-}
-
-#[derive(Component)]
-pub struct UiBackgroundMarker;
-
-fn set_ui_background(
-    mut commands: Commands,
-    backgrounds: Query<
-        (&SceneEntity, &UiBackground, &UiLink),
-        Or<(Changed<UiBackground>, Changed<UiLink>)>,
-    >,
-    mut removed: RemovedComponents<UiBackground>,
-    links: Query<&UiLink>,
-    children: Query<&Children>,
-    prev_backgrounds: Query<Entity, With<UiBackgroundMarker>>,
-    contexts: Query<&RendererSceneContext>,
-    resolver: TextureResolver,
-    mut stretch_uvs: ResMut<Assets<StretchUvMaterial>>,
-) {
-    for ent in removed.read() {
-        let Ok(link) = links.get(ent) else {
-            continue;
-        };
-
-        if let Ok(children) = children.get(link.ui_entity) {
-            for child in children
-                .iter()
-                .filter(|c| prev_backgrounds.get(**c).is_ok())
-            {
-                if let Some(commands) = commands.get_entity(*child) {
-                    commands.despawn_recursive();
-                }
-            }
-        }
-
-        if let Some(mut commands) = commands.get_entity(link.ui_entity) {
-            commands.remove::<BackgroundColor>();
-        }
-    }
-
-    for (scene_ent, background, link) in backgrounds.iter() {
-        let Some(mut commands) = commands.get_entity(link.ui_entity) else {
-            continue;
-        };
-
-        debug!("[{}] set background {:?}", scene_ent.id, background);
-
-        if let Some(texture) = background.texture.as_ref() {
-            let Ok(ctx) = contexts.get(scene_ent.root) else {
-                continue;
-            };
-
-            let image = texture
-                .tex
-                .tex
-                .as_ref()
-                .and_then(|tex| resolver.resolve_texture(ctx, tex).ok());
-
-            let texture_mode = match texture.tex.tex {
-                Some(texture_union::Tex::Texture(_)) => texture.mode,
-                _ => BackgroundTextureMode::stretch_default(),
-            };
-
-            if let Some(image) = image {
-                let image_color = background.color.unwrap_or(Color::WHITE);
-                let image_color = image_color.with_alpha(image_color.alpha() * link.opacity.0);
-                match texture_mode {
-                    BackgroundTextureMode::NineSlices(rect) => {
-                        commands.try_with_children(|c| {
-                            c.spawn((
-                                NodeBundle {
-                                    style: Style {
-                                        position_type: PositionType::Absolute,
-                                        width: Val::Percent(100.0),
-                                        height: Val::Percent(100.0),
-                                        overflow: Overflow::clip(),
-                                        ..Default::default()
-                                    },
-                                    ..Default::default()
-                                },
-                                Ui9Slice {
-                                    image: image.image,
-                                    center_region: rect.into(),
-                                    tint: Some(image_color),
-                                },
-                            ));
-                        });
-                    }
-                    BackgroundTextureMode::Stretch(ref uvs) => {
-                        commands.try_with_children(|c| {
-                            c.spawn(NodeBundle {
-                                style: Style {
-                                    position_type: PositionType::Absolute,
-                                    width: Val::Percent(100.0),
-                                    height: Val::Percent(100.0),
-                                    overflow: Overflow::clip(),
-                                    ..Default::default()
-                                },
-                                ..Default::default()
-                            })
-                            .try_with_children(|c| {
-                                c.spawn((MaterialNodeBundle {
-                                    style: Style {
-                                        position_type: PositionType::Absolute,
-                                        width: Val::Percent(100.0),
-                                        height: Val::Percent(100.0),
-                                        ..Default::default()
-                                    },
-                                    material: stretch_uvs.add(StretchUvMaterial {
-                                        image: image.image.clone(),
-                                        uvs: *uvs,
-                                        color: image_color.to_linear().to_vec4(),
-                                    }),
-                                    ..Default::default()
-                                },));
-                            });
-                        });
-                    }
-                    BackgroundTextureMode::Center => {
-                        commands.try_with_children(|c| {
-                            // make a stretchy grid
-                            c.spawn(NodeBundle {
-                                style: Style {
-                                    position_type: PositionType::Absolute,
-                                    left: Val::Px(0.0),
-                                    right: Val::Px(0.0),
-                                    top: Val::Px(0.0),
-                                    bottom: Val::Px(0.0),
-                                    justify_content: JustifyContent::Center,
-                                    overflow: Overflow::clip(),
-                                    width: Val::Percent(100.0),
-                                    ..Default::default()
-                                },
-                                ..Default::default()
-                            })
-                            .try_with_children(|c| {
-                                c.spacer();
-                                c.spawn(NodeBundle {
-                                    style: Style {
-                                        flex_direction: FlexDirection::Column,
-                                        justify_content: JustifyContent::Center,
-                                        overflow: Overflow::clip(),
-                                        height: Val::Percent(100.0),
-                                        ..Default::default()
-                                    },
-                                    ..Default::default()
-                                })
-                                .try_with_children(|c| {
-                                    c.spacer();
-                                    c.spawn(ImageBundle {
-                                        style: Style {
-                                            overflow: Overflow::clip(),
-                                            ..Default::default()
-                                        },
-                                        image: UiImage {
-                                            color: image_color,
-                                            texture: image.image,
-                                            flip_x: false,
-                                            flip_y: false,
-                                        },
-                                        ..Default::default()
-                                    });
-                                    c.spacer();
-                                });
-                                c.spacer();
-                            });
-                        });
-                    }
-                }
-            } else {
-                warn!("failed to load ui image from content map: {:?}", texture);
-            }
-        } else if let Some(color) = background.color {
-            commands.insert(BackgroundColor(color));
-        }
-    }
-}
-
-pub trait ValAsPx {
-    fn as_px(&self) -> f32;
-}
-
-impl ValAsPx for Val {
-    fn as_px(&self) -> f32 {
-        match self {
-            Val::Px(px) => *px,
-            _ => 0.0,
         }
     }
 }
